New Nav Redirects (#27)
* Revamp user API endpoints, add user hook

Added a hook for fetching our user object for the current authenticated
user. The hook consumes some API endpoints that I changed around a bit.

* Added improved nav logic to pages

Added the redirect and stuff to the auth flow. It's not quite complete
because we don't have applications associated with users. It's a solid
start though and does redirect to profile
info right now.

* Fix logic, rename page